+++ EmployeeLeaveManagementWebAPI/DAL/EmployeeLeaveTransactionHistory.cs
@@ -20,7 +20,7 @@ namespace LMS_WebAPI_DAL
         public System.DateTime ToDate { get; set; }
         public System.DateTime CreatedDate { get; set; }
         public int RefStatus { get; set; }
-        public int NumberOfWorkingDays { get; set; }
+        public double NumberOfWorkingDays { get; set; }
         public int RefLeaveType { get; set; }
         public string EmployeeComment { get; set; }
         public string ManagerComment { get; set; }
