+++ EmployeeLeaveManagementWebAPI/DAL/Repositories/Interfaces/IAddLeaveRepository.cs
@@ -10,7 +10,7 @@ namespace LMS_WebAPI_DAL.Repositories.Interfaces
   public  interface IAddLeaveRepository
     {
         List<string> GetLeaveType();
-        bool InsertEmployeeLeaveDetails(int empId,int leaveType, string fromDate, string toDate, string comments, int workingDays);
+        bool InsertEmployeeLeaveDetails(int empId,int leaveType, string fromDate, string toDate, string comments, double workingDays);
 
         bool SubmitLeaveForApproval(int id);
