+++ EmployeeLeaveManagementWebAPI/Domain/EmployeeDetailsModel.cs
@@ -24,8 +24,8 @@ namespace LMS_WebAPI_Domain
         public string ProjectName { get; set; }
         public String RoleName { get; set; }
         public int? TotalLeaveCount { get; set; }
-        public int TotalApplied { get; set; }
-        public int TotalSpent { get; set; }
+        public double TotalApplied { get; set; }
+        public double TotalSpent { get; set; }
         public string ManagerName { get; set; }
         public string MangerEmail { get; set; }
         public int? ManagerId { get; set; }
