+++ EmployeeLeaveManagementWebAPI/Utils/EmployeeCommon.cs
@@ -22,8 +22,8 @@ namespace LMS_WebAPI_Utils
         public Nullable<System.DateTime> Lastlogin { get; set; }
         public string ProjectName { get; set; }
         public int? TotalLeaveCount { get; set; }
-        public int TotalApplied { get; set; }
-        public int TotalSpent { get; set; }
+        public double TotalApplied { get; set; }
+        public double TotalSpent { get; set; }
         public string ManagerEmailId { get; set; }
         public IDictionary<string,int> MonthlyLeaveReport { get; set; }
     }
